[Refactor] move `SlashKeyboardController` into `BlockInsertPalette`
